Feat: redis 설정 추가
