+++ Common/src/main/java/com/seveneleven/config/TokenProvider.java
@@ -3,7 +3,7 @@ package com.seveneleven.config;
 import com.seveneleven.entity.member.RefreshToken;
 import com.seveneleven.util.security.dto.CustomUserDetails;
 import com.seveneleven.util.security.dto.TokenResponse;
-import com.seveneleven.util.security.repository.RefreshTokenRepositoryImpl;
+import com.seveneleven.util.security.repository.RefreshTokenRepository;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
@@ -11,18 +11,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Component;
 
 import java.security.Key;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.Date;
 import java.util.stream.Collectors;
 
@@ -37,7 +30,7 @@ public class TokenProvider implements InitializingBean {
     private static final String AUTHORITIES_KEY = "auth";
     private final long ACCESS_TOKEN_EXPIRE_TIME;
     private final long REFRESH_TOKEN_EXPIRE_TIME;
-    private final RefreshTokenRepositoryImpl refreshTokenRepositoryImpl;
+    private final RefreshTokenRepository refreshTokenRepositoryImpl;
     private final String secret;
     private Key key;
 
@@ -51,7 +44,7 @@ public class TokenProvider implements InitializingBean {
     public TokenProvider(
             @Value("${jwt.secret}") String secret,
             @Value("${jwt.access-token-validity}") long acessTokenValidityInSeconds,
-            @Value("${jwt.refresh-token-validity}") long refreshTokenValidityInSeconds, RefreshTokenRepositoryImpl refreshTokenRepositoryImpl) {
+            @Value("${jwt.refresh-token-validity}") long refreshTokenValidityInSeconds, RefreshTokenRepository refreshTokenRepositoryImpl) {
         this.secret = secret;
         this.ACCESS_TOKEN_EXPIRE_TIME = acessTokenValidityInSeconds * 1000;
         this.REFRESH_TOKEN_EXPIRE_TIME = refreshTokenValidityInSeconds * 1000;
@@ -133,79 +126,45 @@ public class TokenProvider implements InitializingBean {
                 .compact();
     }
 
-
-    public String generateTokenDto(Authentication authentication) {
-        // 권한들 가져오기
-        String authorities = authentication.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.joining(","));
-
-        long now = (new Date()).getTime();
-
-        // Access Token 생성
-        Date accessTokenExpiresIn = new Date(now + ACCESS_TOKEN_EXPIRE_TIME);
-        String accessToken = Jwts.builder()
-                .setSubject(authentication.getName())       // payload "sub": "name"
-                .claim(AUTHORITIES_KEY, authorities)        // payload "auth": "USER"
-                .setExpiration(accessTokenExpiresIn)        // payload "exp": 151621022 (ex)
-                .signWith(key, SignatureAlgorithm.HS512)    // header "alg": "HS512"
-                .compact();
-
-        // Refresh Token 생성
-        String refreshToken = Jwts.builder()
-                .setExpiration(new Date(now + REFRESH_TOKEN_EXPIRE_TIME))
-                .signWith(key, SignatureAlgorithm.HS512)
-                .compact();
-
-        return "";
-//        return TokenDto.builder()
-//                .grantType(BEARER_TYPE)
-//                .accessToken(accessToken)
-//                .accessTokenExpiresIn(accessTokenExpiresIn.getTime())
-//                .refreshToken(refreshToken)
-//                .build();
-    }
-
-
-    /*
-    * 함수명 : getExpirationFromToken
-    * JWT 만료 시간 추출
-    *
-    * @param token JWT 토큰    
-    * @return 만료 시간 반환
-    * */
-    public LocalDateTime getExpirationFromToken(String token) {
-        Claims claims = Jwts.parser()
-                .setSigningKey(secret)
-                .parseClaimsJws(token)
-                .getBody();
-
-        Date expiration = claims.getExpiration();
-        return expiration.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
-    }
-
-    /**
-     * JWT 토큰을 파싱하여 Authentication 객체를 생성합니다.
-     *
-     * @param token JWT 토큰
-     * @return Authentication 객체
-     */
-    public Authentication getAuthentication(String token) {
-        Claims claims = Jwts.parserBuilder()
-                .setSigningKey(key)
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-
-        Collection<? extends GrantedAuthority> authorities =
-                Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
-                        .map(SimpleGrantedAuthority::new)
-                        .collect(Collectors.toList());
-
-        User principal = new User(claims.getSubject(), "", authorities);
-
-        return new UsernamePasswordAuthenticationToken(principal, token, authorities);
-    }
+//    /*
+//    * 함수명 : getExpirationFromToken
+//    * JWT 만료 시간 추출
+//    *
+//    * @param token JWT 토큰
+//    * @return 만료 시간 반환
+//    * */
+//    public LocalDateTime getExpirationFromToken(String token) {
+//        Claims claims = Jwts.parser()
+//                .setSigningKey(secret)
+//                .parseClaimsJws(token)
+//                .getBody();
+//
+//        Date expiration = claims.getExpiration();
+//        return expiration.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime();
+//    }
+//
+//    /**
+//     * JWT 토큰을 파싱하여 Authentication 객체를 생성합니다.
+//     *
+//     * @param token JWT 토큰
+//     * @return Authentication 객체
+//     */
+//    public Authentication getAuthentication(String token) {
+//        Claims claims = Jwts.parserBuilder()
+//                .setSigningKey(key)
+//                .build()
+//                .parseClaimsJws(token)
+//                .getBody();
+//
+//        Collection<? extends GrantedAuthority> authorities =
+//                Arrays.stream(claims.get(AUTHORITIES_KEY).toString().split(","))
+//                        .map(SimpleGrantedAuthority::new)
+//                        .collect(Collectors.toList());
+//
+//        User principal = new User(claims.getSubject(), "", authorities);
+//
+//        return new UsernamePasswordAuthenticationToken(principal, token, authorities);
+//    }
 
     /**
      * Token에서 login ID 추출
