+++ Common/src/main/java/com/seveneleven/util/security/service/AuthService.java
@@ -1,6 +1,6 @@
-package com.seveneleven.util.security.service;
-
-public interface AuthService {
-    boolean isRefreshTokenExpired(String refreshToken);
-    boolean isValidRefreshToken(String refreshToken);
-}
+//package com.seveneleven.util.security.service;
+//
+//public interface AuthService {
+//    boolean isRefreshTokenExpired(String refreshToken);
+//    boolean isValidRefreshToken(String refreshToken);
+//}
